Remove example lib and clickhouse script from dev project + move code from example to profiling
